text-mining-service/pyproject.toml
[project]
name = "mining-microservice"
version = "0.1.0"
description = "LLM-based microservice for document mining"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.6.0",
    "boto3>=1.34.0",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "lancedb",
    "python-dotenv",
    "langchain",
    "pyjwt",
    "aiohttp",
    "requests",
    "fastapi",
    "uvicorn"
]
